pyproject.toml
[tool.poetry]
name = "elementary-data"
version = "0.14.1"
description = "Data monitoring and lineage"
authors = ["Elementary"]
keywords = ["data", "lineage", "data lineage", "data warehouse", "DWH", "observability", "data monitoring", "data observability", "Snowflake", "BigQuery", "Redshift", "data reliability", "analytics engineering"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://www.elementary-data.com"
repository = "https://github.com/elementary-data/elementary"
documentation = "https://docs.elementary-data.com"

packages = [{include = "elementary"}]

[tool.poetry.dependencies]
python = ">=3.8,<4"
click = ">=7.0,<8.1.4"
pyfiglet = "<1.0.0"
dbt-core = ">=0.20,<2.0.0"
requests = ">=2.28.1,<3.0.0"
beautifulsoup4 = "<5.0.0"
ratelimit = "*"
posthog = "<3.0.0"
boto3 = "<2.0.0"
google-cloud-storage = "<3.0.0"
"ruamel.yaml" = "<1.0.0"
alive-progress = "<=2.3.1"
slack-sdk = ">=3.20.1,<4.0.0"
pytest-parametrization = ">=2022.2.1"

pydantic = "<3.0"
networkx = ">=2.3,<3"
packaging = ">=20.9,<=23.1"
azure-storage-blob = ">=12.11.0"
pymsteams = ">=0.2.2,<1.0.0"
pandas = ">=2.0.0"
tabulate = ">= 0.9.0"

dbt-snowflake = {version = ">=0.20,<2.0.0", optional = true}
dbt-bigquery = {version = ">=0.20,<2.0.0", optional = true}
dbt-redshift = {version = ">=0.20,<2.0.0", optional = true}
dbt-postgres = {version = ">=0.20,<2.0.0", optional = true}
dbt-databricks = {version = ">=0.20,<2.0.0", optional = true}
dbt-spark = {version = ">=0.20,<2.0.0", optional = true}
dbt-athena-community = {version = ">=1.6.3,<2.0.0", optional = true}

[tool.poetry.extras]
snowflake = ["dbt-snowflake"]
bigquery = ["dbt-bigquery"]
redshift = ["dbt-redshift"]
postgres = ["dbt-postgres"]
databricks = ["dbt-databricks"]
spark = ["dbt-spark"]
athena = ["dbt-athena-community"]
all = ["dbt-snowflake", "dbt-bigquery", "dbt-redshift", "dbt-postgres", "dbt-databricks", "dbt-spark"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
edr = "elementary.cli.cli:cli"
